Moves gen test commands to commands.js
